add preconnect handler
