+++ dev_utils/mockServer.ts
@@ -2,16 +2,17 @@ import { createWebStreamPair } from "./webStreamPair.ts";
 import { MqttConn } from "../mqttConn/mqttConn.ts";
 import { Context, MqttServer } from "../server/mod.ts";
 import { handlers } from "./test-handlers.ts";
-import type { IPersistence } from "../persistence/mod.ts";
+import type { MqttServerOptions } from "../server/mod.ts";
 
-export function startMockServer(persistence?: IPersistence): {
+export function startMockServer(serverOpts: MqttServerOptions = { handlers }): {
   mqttConn: MqttConn;
   mqttServer: MqttServer;
 } {
   // start with a fresh clientlist
   Context.clientList.clear();
   // create a new MQTT server
-  const mqttServer = new MqttServer({ handlers, persistence });
+
+  const mqttServer = new MqttServer(serverOpts);
   const { input, output } = createWebStreamPair();
   const mqttConn = new MqttConn({ conn: output });
   mqttServer.serve(input);
@@ -25,23 +26,23 @@ export function addMockClient(mqttServer: MqttServer): MqttConn {
   return mqttConn;
 }
 
-export function startMockServer2(persistence?: IPersistence): {
+export function startMockServer2(serverOpts?: MqttServerOptions): {
   mqttConn1: MqttConn;
   mqttConn2: MqttConn;
   mqttServer: MqttServer;
 } {
-  const { mqttConn: mqttConn1, mqttServer } = startMockServer(persistence);
+  const { mqttConn: mqttConn1, mqttServer } = startMockServer(serverOpts);
   const mqttConn2 = addMockClient(mqttServer);
   return { mqttConn1, mqttConn2, mqttServer };
 }
 
-export function startMockServer3(persistence?: IPersistence): {
+export function startMockServer3(serverOpts?: MqttServerOptions): {
   mqttConn1: MqttConn;
   mqttConn2: MqttConn;
   mqttConn3: MqttConn;
   mqttServer: MqttServer;
 } {
-  const { mqttConn1, mqttConn2, mqttServer } = startMockServer2(persistence);
+  const { mqttConn1, mqttConn2, mqttServer } = startMockServer2(serverOpts);
   const mqttConn3 = addMockClient(mqttServer);
   return { mqttConn1, mqttConn2, mqttConn3, mqttServer };
 }
